Extract method related mekanism to other class
